feat(WebView): Support for IsScrollEnabled
