+++ src/Uno.UI/UI/Xaml/Controls/WebView/Native/Android/NativeWebViewWrapper.Android.cs
@@ -73,6 +73,12 @@ internal class NativeWebViewWrapper : INativeWebView
 
 	public void Reload() => _webView.Reload();
 
+	public void SetScrollingEnabled(bool isScrollingEnabled)
+	{	
+		_webView.HorizontalScrollBarEnabled = isScrollingEnabled;
+		_webView.VerticalScrollBarEnabled = isScrollingEnabled;
+	}
+
 	private void GoToNearestValidHistoryEntry(int direction) =>
 		Enumerable
 			.Repeat(
