+++ src/Uno.UI/UI/Xaml/Controls/WebView/Native/iOSmacOS/INativeWebView.iOSmacOS.cs
@@ -14,6 +14,4 @@ namespace Uno.UI.Xaml.Controls;
 internal partial interface INativeWebView
 {
 	void SetOwner(CoreWebView2 xamlWebView);
-
-	void SetScrollingEnabled(bool isScrollingEnabled);
 }
