+++ src/Uno.UI/UI/Xaml/Controls/WebView/Native/iOSmacOS/WebViewNavigationDelegate.iOSmacOS.cs
@@ -238,176 +238,6 @@ internal class WebViewNavigationDelegate : WKNavigationDelegate
 		}
 	}
 
-	private void OpenUrl(string url)
-	{
-		//TODO: This should be used
-		//TODO: CanGoBack and CanGoForward should be refreshed after navigation completed on both Android and iOS
-		var nsUrl = new NSUrl(url);
-		//Opens the specified URL, launching the app that's registered to handle the scheme.
-#if __IOS__
-		UIApplication.SharedApplication.OpenUrl(nsUrl);
-#else
-		NSWorkspace.SharedWorkspace.OpenUrl(nsUrl);
-#endif
-	}
-
-	internal void OnNavigationStarting(WebViewNavigationStartingEventArgs args)
-	{
-		if (args.Uri == null)
-		{
-			//_owner case should not happen when navigating normally using http requests.
-			//_owner is to stop a scenario where the webview is initialized without having a source
-			args.Cancel = true;
-			return;
-		}
-
-		if (args.Uri.Scheme.Equals(Uri.UriSchemeMailto, StringComparison.OrdinalIgnoreCase))
-		{
-#if __IOS__
-			ParseUriAndLauchMailto(args.Uri);
-#else
-						NSWorkspace.SharedWorkspace.OpenUrl(new NSUrl(args.Uri.ToString()));
-#endif
-			args.Cancel = true;
-			return;
-		}
-
-		//TODO:MZ:
-		//NavigationStarting?.Invoke(_owner, args);
-	}
-
-#if __IOS__
-	private void ParseUriAndLauchMailto(Uri mailtoUri)
-	{
-		_ = Uno.UI.Dispatching.CoreDispatcher.Main.RunAsync(
-			Uno.UI.Dispatching.CoreDispatcherPriority.Normal,
-			async (ct) =>
-			{
-				try
-				{
-					var subject = "";
-					var body = "";
-					var cc = new[] { "" };
-					var bcc = new[] { "" };
-
-					var recipients = mailtoUri.AbsoluteUri.Split(new[] { ':' })[1].Split(new[] { '?' })[0].Split(new[] { ',' });
-					var parameters = mailtoUri.Query.Split(new[] { '?' });
-
-					parameters = parameters.Length > 1 ?
-									parameters[1].Split(new[] { '&' }) :
-									Array.Empty<string>();
-
-					foreach (string param in parameters)
-					{
-						var keyValue = param.Split(new[] { '=' });
-						var key = keyValue[0];
-						var value = keyValue[1];
-
-						switch (key)
-						{
-							case "subject":
-								subject = value;
-								break;
-							case "to":
-								recipients.Concat(value);
-								break;
-							case "body":
-								body = value;
-								break;
-							case "cc":
-								cc.Concat(value);
-								break;
-							case "bcc":
-								bcc.Concat(value);
-								break;
-							default:
-								break;
-						}
-					}
-
-					await LaunchMailto(ct, subject, body, recipients, cc, bcc);
-				}
-
-				catch (Exception e)
-				{
-					if (this.Log().IsEnabled(Uno.Foundation.Logging.LogLevel.Error))
-					{
-						this.Log().Error("Unable to launch mailto", e);
-					}
-				}
-			});
-	}
-
-	public
-#if !__MACCATALYST__
-	async
-#endif
-	Task LaunchMailto(CancellationToken ct, string subject = null, string body = null, string[] to = null, string[] cc = null, string[] bcc = null)
-	{
-#if !__MACCATALYST__  // catalyst https://github.com/xamarin/xamarin-macios/issues/13935
-		if (!MFMailComposeViewController.CanSendMail)
-		{
-			return;
-		}
-
-		var mailController = new MFMailComposeViewController();
-
-		mailController.SetToRecipients(to);
-		mailController.SetSubject(subject);
-		mailController.SetMessageBody(body, false);
-		mailController.SetCcRecipients(cc);
-		mailController.SetBccRecipients(bcc);
-
-		var finished = new TaskCompletionSource<object>();
-		var handler = new EventHandler<MFComposeResultEventArgs>((snd, args) => finished.TrySetResult(args));
-
-		try
-		{
-			mailController.Finished += handler;
-			UIApplication.SharedApplication.KeyWindow.RootViewController.PresentViewController(mailController, true, null);
-
-			using (ct.Register(() => finished.TrySetCanceled()))
-			{
-				await finished.Task;
-			}
-		}
-		finally
-		{
-			await CoreDispatcher
-				.Main
-				.RunAsync(CoreDispatcherPriority.High, () =>
-				{
-					mailController.Finished -= handler;
-					mailController.DismissViewController(true, null);
-				})
-				.AsTask(CancellationToken.None);
-		}
-#else
-		return Task.CompletedTask;
-#endif
-	}
-#endif
-
-	internal void OnNewWindowRequested(WebViewNewWindowRequestedEventArgs args)
-	{
-		//TODO:MZ:
-		//NewWindowRequested?.Invoke(_owner, args);
-	}
-
-	internal void OnNavigationFailed(WebViewNavigationFailedEventArgs args)
-	{
-		//TODO:MZ:
-		//NavigationFailed?.Invoke(_owner, args);
-	}
-
-
 	[Obsolete("https://github.com/unoplatform/uno/pull/1591")]
 	internal static bool MustUseWebKitWebView() => true;
-
-	private void OnScrollEnabledChangedPartial(bool scrollingEnabled)
-	{
-		//TODO:MZ:
-		//_nativeWebView.SetScrollingEnabled(scrollingEnabled);
-	}
-
 }
